+++ client/src/app/routes/users/users-tasks-dialog/users-tasks-dialog.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import saveAs from 'file-saver';
 import { TodosService } from '../../../services/todos.service';
-import { ToDoTask, UserTaskFile } from '../../../shared/models/todo';
+import { UserTask, UserTaskFile } from '../../../shared/models/todo';
 
 @Component({
   selector: 'app-users-tasks-dialog',
@@ -14,7 +14,7 @@ import { ToDoTask, UserTaskFile } from '../../../shared/models/todo';
 export class UsersTasksDialogComponent implements OnInit {
   @Input() userId: number | undefined;
   @Output() tasksViewClosed = new EventEmitter();
-  todos: ToDoTask[] = [];
+  todos: UserTask[] = [];
 
   constructor(private todosService: TodosService) {}
   ngOnInit(): void {
@@ -27,9 +27,9 @@ export class UsersTasksDialogComponent implements OnInit {
 
   private loadTodos() {
     {
-      this.todosService.getUserTodos(this.userId!, 1, 5).subscribe({
+      this.todosService.getUserTasks(this.userId!).subscribe({
         next: (res) => {
-          this.todos = res.tasks;
+          this.todos = res;
         },
       });
     }