+++ client/changes.md
@@ -92,13 +92,13 @@
 - fix sharing
 - fix roles updating
 
-## 7-2-2024
+### 7-2-2024
 
 - show shared with, allow editing shared with
 - get all users and cache them. Stop getting users on share click
 - add loading indicators --
-- users must have at least one role
-- show only user tasks in profiles page, not tasks shared with
+- users must have at least one role --
+- show only user tasks in profiles page, not tasks shared with --
 - allow adding or deleting files on todo update
 - FIX UPLOAD BUTTON --
 - on new todo added, move to top of the list --